# openapi/tensorzero-gateway-external.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TensorZero Gateway API",
    "description": "TensorZero Gateway endpoints.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "2026.3.2"
  },
  "paths": {
...
    "/v1/inferences/get_inferences": {
      "post": {
        "tags": [
          "inferences"
        ],
        "summary": "Handler for the POST `/v1/inferences/get_inferences` endpoint.\nRetrieves specific inferences by their IDs.",
        "operationId": "get_inferences_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetInferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Get stored inferences by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetInferencesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TensorZeroErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TensorZeroErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/inferences/list_inferences": {
      "post": {
        "tags": [
          "inferences"
        ],
        "summary": "Handler for the POST `/v1/inferences/list_inferences` endpoint.\nLists inferences with optional filtering, pagination, and sorting.",
        "operationId": "list_inferences_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListInferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "List stored inferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetInferencesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TensorZeroErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TensorZeroErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "GetInferencesRequest": {
        "type": "object",
        "description": "Request to get specific inferences by their IDs.\nUsed by the `POST /v1/inferences/get_inferences` endpoint.",
        "required": [
          "ids"
        ],
        "properties": {
          "function_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional function name to filter by.\nIncluding this improves query performance since `function_name` is the first column\nin the ClickHouse primary key."
          },
          "ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The IDs of the inferences to retrieve. Required."
          },
          "output_source": {
            "$ref": "#/components/schemas/InferenceOutputSource",
            "description": "Source of the inference output.\nDetermines whether to return the original inference output or demonstration feedback\n(manually-curated output) if available.\nDefaults to `Inference` if not specified."
          }
        }
      },
      "GetInferencesResponse": {
        "type": "object",
        "description": "Response containing the requested inferences.",
        "required": [
          "inferences"
        ],
        "properties": {
          "inferences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StoredInference"
            },
            "description": "The retrieved inferences."
          }
        }
      },
...
      "InferenceOutputSource": {
        "type": "string",
        "description": "Source of an inference output when querying inferences. Users can choose this because there may be\ndemonstration feedback (manually-curated output) for the inference that should be preferred.",
        "enum": [
          "none",
          "inference",
          "demonstration"
        ]
      },
...
      "ListInferencesRequest": {
        "type": "object",
        "description": "Request to list inferences with pagination and filters.\nUsed by the `POST /v1/inferences/list_inferences` endpoint.",
        "properties": {
          "after": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Optional inference ID to paginate after (exclusive).\nReturns inferences with IDs after this one (later in time).\nCannot be used together with `before` or `offset`."
          },
          "before": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Optional inference ID to paginate before (exclusive).\nReturns inferences with IDs before this one (earlier in time).\nCannot be used together with `after` or `offset`."
          },
          "episode_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Optional episode ID to filter inferences by.\nIf provided, only inferences from this episode will be returned."
          },
          "filters": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/InferenceFilter",
                "description": "Optional filter to apply when querying inferences.\nSupports filtering by metrics, tags, time, and logical combinations (AND/OR/NOT)."
              }
            ]
          },
          "function_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional function name to filter inferences by.\nIf provided, only inferences from this function will be returned."
          },
          "limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The maximum number of inferences to return.\nDefaults to 20.",
            "minimum": 0
          },
          "offset": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The number of inferences to skip before starting to return results.\nDefaults to 0.",
            "minimum": 0
          },
          "order_by": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/OrderBy"
            },
            "description": "Optional ordering criteria for the results.\nSupports multiple sort criteria (e.g., sort by timestamp then by metric)."
          },
          "output_source": {
            "$ref": "#/components/schemas/InferenceOutputSource",
            "description": "Source of the inference output. Determines whether to return the original\ninference output or demonstration feedback (manually-curated output) if available.\nDefaults to `Inference` if not specified."
          },
          "search_query_experimental": {
            "type": [
              "string",
              "null"
            ],
            "description": "Text query to filter. Case-insensitive substring search over the inferences' input and output.\n\nTHIS FEATURE IS EXPERIMENTAL, and we may change or remove it at any time.\nWe recommend against depending on this feature for critical use cases.\n\nImportant limitations:\n- This requires an exact substring match; we do not tokenize this query string.\n- This doesn't search for any content in the template itself.\n- Quality is based on term frequency > 0, without any relevance scoring.\n- There are no performance guarantees (it's best effort only). Today, with no other\n  filters, it will perform a full table scan, which may be extremely slow depending\n  on the data volume."
          },
          "variant_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional variant name to filter inferences by.\nIf provided, only inferences from this variant will be returned."
          }
        }
      },
...
      "OrderBy": {
        "allOf": [
          {
            "$ref": "#/components/schemas/OrderByTerm",
            "description": "The property to order by."
          },
          {
            "type": "object",
            "required": [
              "direction"
            ],
            "properties": {
              "direction": {
                "$ref": "#/components/schemas/OrderDirection",
                "description": "The ordering direction."
              }
            }
          }
        ],
        "description": "Order by clauses for querying inferences."
      },
...
      "OrderDirection": {
        "type": "string",
        "description": "The ordering direction.",
        "enum": [
          "ascending",
          "descending"
        ]
      },
...
      "StoredChatInference": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DynamicToolParams"
          },
          {
            "type": "object",
            "required": [
              "function_name",
              "variant_name",
              "timestamp",
              "episode_id",
              "inference_id"
            ],
            "properties": {
              "dispreferred_outputs": {
                "type": "array",
                "items": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContentBlockChatOutput"
                  }
                }
              },
              "episode_id": {
                "type": "string",
                "format": "uuid"
              },
              "extra_body": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/UnfilteredInferenceExtraBody"
                  }
                ]
              },
              "function_name": {
                "type": "string"
              },
              "inference_id": {
                "type": "string",
                "format": "uuid"
              },
              "inference_params": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/InferenceParams"
                  }
                ]
              },
              "input": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/StoredInput"
                  }
                ]
              },
              "output": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "$ref": "#/components/schemas/ContentBlockChatOutput"
                }
              },
              "processing_time_ms": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64",
                "minimum": 0
              },
              "snapshot_hash": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "tags": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                },
                "propertyNames": {
                  "type": "string"
                }
              },
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "ttft_ms": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int64",
                "minimum": 0
              },
              "variant_name": {
                "type": "string"
              }
            }
          }
        ],
        "description": "Wire variant of StoredChatInference for API responses with Python/TypeScript bindings"
      },
...
      "StoredInference": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/StoredChatInference"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "chat"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/StoredJsonInference"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "json"
                    ]
                  }
                }
              }
            ]
          }
        ],
        "description": "Wire variant of StoredInference for API responses with Python/TypeScript bindings\nThis one should be used in all public interfaces"
      },
...
      "TensorZeroErrorResponse": {
        "type": "object",
        "description": "Stable baseline error envelope for generated SDKs.\n\nWe intentionally document only the top-level `error` string for now.\nInternal structured error payloads remain an implementation detail.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "UnfilteredInferenceExtraBody": {
        "type": "object",
        "description": "The 'InferenceExtraBody' options provided directly in an inference request.\nThese have not yet been filtered by variant name",
        "required": [
          "extra_body"
        ],
        "properties": {
          "extra_body": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExtraBody"
            }
          }
        }
      },
...
    }
  }
}